src/gemini/requirements.txt
flask
Flask-RESTful
flask-cors
pyyaml
pandas
uwsgi